requirements.txt
asgiref==3.7.2
Django==4.2.4
sqlparse==0.4.4
gunicorn
django-heroku
dj-database-url
django-allauth
django-storages
boto3
upload
python-magic==0.4.24
django-bootstrap-v5
